+++ scripts/config-phpcs-installed-paths.sh
@@ -2,6 +2,6 @@
 
 set -euo pipefail
 
-vendor/bin/phpcs --config-set installed_paths "../../phpcsstandards/phpcsutils,../../phpcsstandards/phpcsextra,../../automattic/vipwpcs,../../phpcompatibility/php-compatibility,../../sirbrillig/phpcs-variable-analysis,../../publishpress/publishpress-phpcs-standards/standards,../../wp-coding-standards/wpcs" > /dev/null 2>&1
+vendor/bin/phpcs --config-set installed_paths "../../phpcsstandards/phpcsutils,../../phpcsstandards/phpcsextra,../../automattic/vipwpcs,../../phpcompatibility/php-compatibility,../../sirbrillig/phpcs-variable-analysis,../../publishpress/publishpress-phpcs-standards/standards,../../wp-coding-standards/wpcs,../../wordpress/plugin-check/phpcs-sniffs" > /dev/null 2>&1
 
 echo "PHPCS installed paths configured"